+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/VerifiableCredentialSdJwt.kt
@@ -1,11 +1,14 @@
 package at.asitplus.wallet.lib.data
 
 import at.asitplus.KmmResult.Companion.wrap
+import at.asitplus.signum.indispensable.josef.ConfirmationClaim
 import at.asitplus.signum.indispensable.josef.JsonWebKey
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.decodeFromJsonElement
 
 /**
  * SD-JWT representation of a [VerifiableCredential].
@@ -74,7 +77,7 @@ data class VerifiableCredentialSdJwt(
      * See (I-D.looker-oauth-jwt-cwt-status-list) for more information.
      */
     @SerialName("status")
-    // TODO Implement correct draft
+    // TODO Implement correct draft: draft-ietf-oauth-status-list-05
     val credentialStatus: CredentialStatus? = null,
 
     @SerialName("_sd_alg")
@@ -88,8 +91,20 @@ data class VerifiableCredentialSdJwt(
      * the key identified in this claim.
      */
     @SerialName("cnf")
-    val confirmationKey: JsonWebKey? = null,
+    // Should be [ConfirmationClaim], but was [JsonWebKey] in our previous implementation...
+    val cnfElement: JsonElement? = null,
 ) {
+
+    @Deprecated("Use confirmationClaim", replaceWith = ReplaceWith("confirmationClaim.jsonWebKey"))
+    val confirmationKey: JsonWebKey?
+        get() = cnfElement?.let {
+            runCatching { vckJsonSerializer.decodeFromJsonElement<JsonWebKey>(it) }.getOrNull()
+        }
+
+    val confirmationClaim: ConfirmationClaim?
+        get() = cnfElement?.let {
+            runCatching { vckJsonSerializer.decodeFromJsonElement<ConfirmationClaim>(it) }.getOrNull()
+        }
 
     fun serialize() = vckJsonSerializer.encodeToString(this)