+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Release 5.1.0:
  - SD-JWT:
    - Pass around decoded data with `SdJwtSigned` in several result classes like `VerifyPresentationResult.SuccessSdJwt`
    - Rename `disclosures` to `reconstructed` in several result classes like `AuthnResponseResult.SuccessSdJwt`
+   - Correctly implement confirmation claim in `VerifiableCredentialSdJwt`, migrating from `JsonWebKey` to `ConfirmationClaim`
 
 Release 5.0.1:
  - Update JsonPath4K to 2.4.0